requirements.txt
faiss-cpu
fastapi
pydantic
openai
langchain
langchain-openai
langchain-huggingface
langchain-community
sentence-transformers
torch
uvicorn
